// frontend/generated/updateUsername.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: updateUsername
// ====================================================

export interface updateUsername {
  /**
   * Update Username
   */
  updateUsername: string;
}

export interface updateUsernameVariables {
  username: string;
}